# src/models/index.js
const User = require('./User');
const OTP = require('./OTP');
const ApiKey = require('./ApiKey');
const Message = require('./Message');
const SystemInfo = require('./SystemInfo');
const Banner = require('./Banner');

module.exports = {
  User,
  OTP,
  ApiKey,
  Message,
  SystemInfo,
  Banner
}; 