# swagger.yaml
openapi: 3.0.0
info:
  title: HamrahMes API
  description: API documentation for HamrahMes backend services
  version: 1.0.0
  contact:
    email: support@hamrahmes.com
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message
...
    Message:
      type: object
      properties:
        _id:
          type: string
          example: '60d21b4667d0d8992e610c85'
        sender:
          type: string
          example: 'Tax System'
        recipient:
          type: string
          example: '09123456789'
        content:
          type: string
          example: 'Your tax payment is due next week'
        timestamp:
          type: string
          format: date-time
        senderImage:
          type: string
          example: 'http://localhost:3000/download/abc123.jpg'
...
paths:
...
  /api/messages:
    get:
      summary: Get user messages
      description: Get paginated list of messages for the authenticated user
      tags:
        - Messages
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Items per page
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 28
                      page:
                        type: integer
                        example: 1
                      pages:
                        type: integer
                        example: 3
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /download/{filename}:
    get:
      summary: Download file
      description: Download a file by filename (images, etc.)
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
          description: The name of the file to download
      responses:
        '200':
          description: File downloaded successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Download failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'